+++ include/memory/memory.h
@@ -32,5 +32,5 @@
 #define KILOBYTES (BYTES * 1024)
 #define MEGABYTES (KILOBYTES * 1024)
 
-void format_bytes(double bytes, char *buff, size_t n);
+void format_bytes(double bytes, char *buff, unsigned int n);
 void clear(void *p, size_t s);
